Renamed classes related to CPS enabled org's pagination feature.
