+++ src/Api/Management/Controller/AppByOwnerControllerInterface.php
@@ -19,9 +19,9 @@
 namespace Apigee\Edge\Api\Management\Controller;
 
 use Apigee\Edge\Controller\CpsLimitEntityControllerInterface;
-use Apigee\Edge\Controller\CpsListingEntityControllerInterface;
 use Apigee\Edge\Controller\EntityControllerInterface;
 use Apigee\Edge\Controller\EntityCrudOperationsControllerInterface;
+use Apigee\Edge\Controller\PaginationEntityListingControllerInterface;
 use Apigee\Edge\Controller\StatusAwareEntityControllerInterface;
 
 /**
@@ -32,7 +32,7 @@ interface AppByOwnerControllerInterface extends
     EntityControllerInterface,
     EntityCrudOperationsControllerInterface,
     CpsLimitEntityControllerInterface,
-    CpsListingEntityControllerInterface,
+    PaginationEntityListingControllerInterface,
     StatusAwareEntityControllerInterface
 {
     /**
