+++ src/Api/Management/Controller/DeveloperAppController.php
@@ -20,8 +20,8 @@ namespace Apigee\Edge\Api\Management\Controller;
 
 use Apigee\Edge\Api\Management\Entity\DeveloperApp;
 use Apigee\Edge\ClientInterface;
-use Apigee\Edge\Controller\CpsListingEntityControllerTrait;
 use Apigee\Edge\Controller\EntityCrudOperationsControllerTrait;
+use Apigee\Edge\Controller\PaginationEntityListingControllerTrait;
 use Apigee\Edge\Controller\StatusAwareEntityControllerTrait;
 use Psr\Http\Message\UriInterface;
 
@@ -33,7 +33,7 @@ class DeveloperAppController extends AppByOwnerController implements DeveloperAp
     use AttributesAwareEntityControllerTrait;
     use AppControllerTrait;
     use EntityCrudOperationsControllerTrait;
-    use CpsListingEntityControllerTrait;
+    use PaginationEntityListingControllerTrait;
     use StatusAwareEntityControllerTrait;
 
     /** @var string Developer email or id. */
