+++ src/Api/Management/Controller/DeveloperController.php
@@ -23,8 +23,8 @@ use Apigee\Edge\Api\Management\Entity\DeveloperInterface;
 use Apigee\Edge\Api\Management\Exception\DeveloperNotFoundException;
 use Apigee\Edge\ClientInterface;
 use Apigee\Edge\Controller\CpsLimitEntityController;
-use Apigee\Edge\Controller\CpsListingEntityControllerTrait;
 use Apigee\Edge\Controller\EntityCrudOperationsControllerTrait;
+use Apigee\Edge\Controller\PaginationEntityListingControllerTrait;
 use Apigee\Edge\Controller\StatusAwareEntityControllerTrait;
 use Apigee\Edge\Denormalizer\AttributesPropertyDenormalizer;
 use Psr\Http\Message\UriInterface;
@@ -35,8 +35,8 @@ use Psr\Http\Message\UriInterface;
 class DeveloperController extends CpsLimitEntityController implements DeveloperControllerInterface
 {
     use AttributesAwareEntityControllerTrait;
-    use CpsListingEntityControllerTrait;
     use EntityCrudOperationsControllerTrait;
+    use PaginationEntityListingControllerTrait;
     use StatusAwareEntityControllerTrait;
 
     /**
