+++ src/Api/Management/Controller/DeveloperControllerInterface.php
@@ -20,9 +20,9 @@ namespace Apigee\Edge\Api\Management\Controller;
 
 use Apigee\Edge\Api\Management\Entity\DeveloperInterface;
 use Apigee\Edge\Controller\CpsLimitEntityControllerInterface;
-use Apigee\Edge\Controller\CpsListingEntityControllerInterface;
 use Apigee\Edge\Controller\EntityControllerInterface;
 use Apigee\Edge\Controller\EntityCrudOperationsControllerInterface;
+use Apigee\Edge\Controller\PaginationEntityListingControllerInterface;
 use Apigee\Edge\Controller\StatusAwareEntityControllerInterface;
 
 /**
@@ -33,9 +33,9 @@ use Apigee\Edge\Controller\StatusAwareEntityControllerInterface;
 interface DeveloperControllerInterface extends
     AttributesAwareEntityControllerInterface,
     CpsLimitEntityControllerInterface,
-    CpsListingEntityControllerInterface,
     EntityControllerInterface,
     EntityCrudOperationsControllerInterface,
+    PaginationEntityListingControllerInterface,
     StatusAwareEntityControllerInterface
 {
     /**
