+++ src/Api/Management/Controller/OrganizationController.php
@@ -22,7 +22,7 @@ use Apigee\Edge\Api\Management\Entity\Organization;
 use Apigee\Edge\ClientInterface;
 use Apigee\Edge\Controller\AbstractEntityController;
 use Apigee\Edge\Controller\EntityCrudOperationsControllerTrait;
-use Apigee\Edge\Controller\NonCpsListingEntityControllerTrait;
+use Apigee\Edge\Controller\NoPaginationEntityListingControllerTrait;
 use Apigee\Edge\Denormalizer\PropertiesPropertyDenormalizer;
 use Apigee\Edge\Normalizer\PropertiesPropertyNormalizer;
 use Psr\Http\Message\UriInterface;
@@ -33,7 +33,7 @@ use Psr\Http\Message\UriInterface;
 class OrganizationController extends AbstractEntityController implements OrganizationControllerInterface
 {
     use EntityCrudOperationsControllerTrait;
-    use NonCpsListingEntityControllerTrait;
+    use NoPaginationEntityListingControllerTrait;
 
     /**
      * OrganizationController constructor.
